# Conflicts:
#	src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/killaura/features/KillAuraFailSwing.kt

+++ src/main/kotlin/net/ccbluex/liquidbounce/deeplearn/data/CombatSample.kt
@@ -90,7 +90,7 @@ data class CombatSample(
             targetDiff.horizontalDistance().toFloat() + playerDiff.horizontalDistance().toFloat(),
 
             // Distance
-            distance.toFloat()
+            distance
         )
 
     val asOutput
